Refactor solution to enforce strict n-tier architecture
This commit corrects several architectural violations to ensure the solution strictly adheres to n-tier principles.

Key changes:
- **Composition Root:** Moved dependency injection setup from the Presentation layer to a new `ServiceFactory` in the BusinessLogic layer. This removes the UI layer's knowledge of data access components.
- **Dependency Fixes:**
  - Removed the direct project reference from `Presentation` to `DataAccess`.
  - Removed the incorrect project reference from `Session` to `BusinessLogic`.
- **Project Cleanup:**
  - Deleted the unused `Common` project and removed all references to it from the solution and other projects.
- **Documentation:**
  - Updated `docs/architecture.md` to accurately reflect the new, correct architecture.

All projects now compile successfully, and existing unit tests pass, confirming the refactoring was successful.
