+++ UserManagementSystem.sln
@@ -7,8 +7,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Presentation", "src\Present
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BusinessLogic", "src\BusinessLogic\BusinessLogic.csproj", "{EDCCA508-1F32-EA37-033F-4FE310356C7E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "src\Common\Common.csproj", "{3F3044DC-A9F8-DE16-79DD-4A0C1649CD06}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataAccess", "src\DataAccess\DataAccess.csproj", "{43EFCA51-8C4A-CDE1-18EA-D7FF1C3FA2D5}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services", "src\Services\Services.csproj", "{23FEB2A4-D9B0-33A5-FA86-CB5942E641D0}"
@@ -35,10 +33,6 @@ Global
 		{EDCCA508-1F32-EA37-033F-4FE310356C7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EDCCA508-1F32-EA37-033F-4FE310356C7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EDCCA508-1F32-EA37-033F-4FE310356C7E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3F3044DC-A9F8-DE16-79DD-4A0C1649CD06}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F3044DC-A9F8-DE16-79DD-4A0C1649CD06}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F3044DC-A9F8-DE16-79DD-4A0C1649CD06}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F3044DC-A9F8-DE16-79DD-4A0C1649CD06}.Release|Any CPU.Build.0 = Release|Any CPU
 		{43EFCA51-8C4A-CDE1-18EA-D7FF1C3FA2D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{43EFCA51-8C4A-CDE1-18EA-D7FF1C3FA2D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{43EFCA51-8C4A-CDE1-18EA-D7FF1C3FA2D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
