+++ src/Presentation/Presentation.csproj
@@ -11,7 +11,6 @@
   <ItemGroup>
     <!-- Referencias a capas internas -->
     <ProjectReference Include="..\BusinessLogic\BusinessLogic.csproj" />
-    <ProjectReference Include="..\DataAccess\DataAccess.csproj" />
     <ProjectReference Include="..\Session\Session.csproj" />
   </ItemGroup>
